banner CRUD Completed

+++ src/main/java/com/zerobase/fastlms/banner/repository/BannerRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface BannerRepository extends JpaRepository<BannerInfo, Integer> {
+public interface BannerRepository extends JpaRepository<BannerInfo, Long> {
     
 }
